增加entity从数据库删除功能. def destroy( self, deleteFromDB, writeToDB ) This method destroys the base part of this entity. It can only be called when there is no cell entity. To destroy the cell entity, call Base.destroyCellEntity.
It may be convenient to call self.destroy in the onLoseCell callback. This causes the base entity to be destroyed whenever the cell entity is destroyed.

This method has two keyword argument:
Parameters: deleteFromDB  If True, the entry in the database associated with this entity is deleted. This defaults to False.
writeToDB  If True, the properties associated with this entity are written to the database. This only occurs if this entity was read from the database or has previously been written with Base.writeToDB. This defaults to True but is ignored if deleteFromDB is True.