fixed:https://github.com/kbengine/kbengine/issues/207 entity.py被删除，但是pyc还存在，服务端启动提示无法导入模块
Python does not check the cache in two circumstances. First, it always recompiles and does not store the result for the module that’s loaded directly from the command line. Second, it does not check the cache if there is no source module. To support a non-source (compiled only) distribution, the compiled module must be in the source directory, and there must not be a source module.